Move text window class to separate file
